A fully automatic vacuum cleaner based on ARM cortex M4
Autovac is an aid for development in consumer electronics. It is used for automated home
vacuuming. It vacuums the house or any given area without any human intervention. It is
an immense help for busy families, elderly and handicapped people. It is convenient and
saves time. It just needs to turn on and leave for certain hours for it to clean the entire space.
Ultrasonic sensors help autovac to navigate avoiding the obstacles, it also helps it to detect
translucent or transparent objects. It can be smarter if a cliff sensor is added to it, so that it
will detect prevent any fall from happening.
The objective of this project is to design a working model of an autonomous vacuum
cleaner from scratch, which can detect and avoid all the obstacles and vacuums the house
simultaneously. The vacuum cleaner is made from high power DC motor and reusable
plastic. It has a simple setup and easy maintenance.